pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "todo4-hermes-plugin"
version = "0.1.0"
description = "Hermes Agent plugin that onboards Todo4 via chat and ships skills to work with Todo4."
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Panit Wechasil" }]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "responses>=0.23",
]

[project.entry-points."hermes_agent.plugins"]
todo4 = "todo4_hermes"

[tool.setuptools.packages.find]
where = ["."]
include = ["todo4_hermes*"]

[tool.setuptools.package-data]
todo4_hermes = ["skills/*.md"]
